Give eslintrc file ending
