--- .eslintrc
@@ -1,10 +1,0 @@
-{
-  "extends": "airbnb",
-  "env": {
-    "node": true,
-    "mocha": true
-  },
-  "rules": {
-    "comma-dangle": "off"
-  }
-}
