--- sample/.npmrc
@@ -1,1 +1,0 @@
-@sap:registry=https://npm.sap.com
