--- sample/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "sap-leonardo-iot-sdk-sample",
-  "version": "1.0.0",
-  "description": "Samples for usage of SAP Leonardo IoT SDK",
-  "scripts": {
-    "start": "node --inspect server.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "engines": {
-    "node": "^8.6.0"
-  },
-  "dependencies": {
-    "sap-leonardo-iot-sdk": "github:SAP/leonardo-iot-sdk-nodejs#v0.1.4",
-    "async-mqtt": "2.3.0",
-    "express": "^4.17.1"
-  },
-  "devDependencies": {
-    "mocha": "^6.1.4"
-  },
-  "author": "SAP SE",
-  "license": "Apache-2.0"
-}
