+++ samples/.npmrc
@@ -1,0 +1,1 @@
+@sap:registry=https://npm.sap.com
